refactor: optimize project structure for scalability
- Move database layer to src/lib/db/ for shared access
- Organize Kanban components in src/components/kanban/
- Separate layout components in src/components/layout/
- Update all import paths across codebase and tests
- Preserve test data in src/data/ for test reuse
- Maintain all functionality while improving organization
- All tests passing (162/162) after restructuring
