+++ src/components/__tests__/kanban-column.test.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
-import { KanbanColumn } from "../kanban-column";
+import { KanbanColumn } from "../kanban/kanban-column";
 import { Task } from "@/types/task";
 
 // Mock the TaskCard component
-jest.mock("../task-card", () => ({
+jest.mock("../kanban/task-card", () => ({
   TaskCard: ({ task }: { task: Task }) => (
     <div data-testid={`task-card-${task.id}`}>
       <h3>{task.title}</h3>
