+++ src/components/kanban/droppable-column.tsx
